change db, added bot client, jdbc processing etc

+++ link-parser/src/main/java/ru/tinkoff/edu/java/linkparser/linkStructures/StackOverflowResult.java
@@ -1,4 +1,4 @@
 package ru.tinkoff.edu.java.linkparser.linkStructures;
 
-public record StackOverflowResult(Integer id) implements Result{
+public record StackOverflowResult(Long id) implements Result{
 }
